fix(api): refactor conversion to another file
Ref. metriport/metriport-internal#1158
